Create create database company
This SQL script will create the database called Company in your MySQL or any other DBMS framework. 
It also builds all the database objects, in accordance with the following book:

Book information
Introduction to Databse systems
  by 
Ramez Elmasri and Shamkant Navate

This is a fairly good database and is handy if you are following this book, because a lot of examples are based 
out of this database design. So you can practice off it.

Page 71-72 of the book explains this database.