Create Library application MySQL schema
Built a PhP application based on this database. This software is built from a Librarians' perspective. It supports multiple library branches.

This database schema defines the structure for a Library, where one can:
1.Open an account
2.Check out new books
3.Check in old books
4.Check fine (if book is being returned after 2 weeks of loan time)
5.Pay fines (if applicable)

The system intelligently checks several aspects to ensure that no duplication in member data is formed.

Additionally in case a member is trying to loan a particular book from a certain library branch, and the book is not available at that particular branch; then the system will inform if the book is available at any other branch, and name that branch.